+++ src/terminal/sys.rs
@@ -1,7 +1,10 @@
 #[cfg(unix)]
-pub(crate) use self::unix::{exit, get_terminal_size};
+pub use self::unix::{exit, get_terminal_size};
+
 #[cfg(windows)]
-pub(crate) use self::winapi::{exit, get_terminal_size};
+pub(crate) use self::winapi::{clear, scroll_down, scroll_up, set_size};
+#[cfg(windows)]
+pub use self::winapi::{exit, get_terminal_size};
 
 #[cfg(windows)]
 pub(crate) mod winapi;
